Rework slot machine animation + UI color scheme
- SlotMachine: real reel scroll animation with acceleration/deceleration
  wheels stop sequentially (0.8s / 1.4s / 2.0s delays)
- Multiplier: average of 3 rolls mapped to 1.0-3.0, one decimal place
- index.css: warm paper-white palette matching reference screenshot
- Wins calendar: larger rounded corners, filled bg for days with records
- tailwind.config: extended border-radius scale

+++ backend/routers/bonus.py
@@ -9,7 +9,7 @@ router = APIRouter(prefix="/bonus", tags=["bonus"])
 class DailyBonus(BaseModel):
     date: str        # YYYY-MM-DD
     rolls: list[int] # 三个数字
-    multiplier: int  # 总倍数
+    multiplier: float  # 倍数（1.0-3.0）
 
 
 @router.get("/today", response_model=DailyBonus | None)
